+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+### 0.16.0
+### Features
+* `ReplicaBroker` class added to simulation to replicate order fills based on real broker data
+### Improvements
+* `print` replaced with `logging` for zerodha broker
+
 ### 0.15.0
 ### Features
 * New broker `Neo` added
+
 ### 0.14.2
 Improvements to `FakeBroker` module
 ### Features